+++ src/js/output.js
@@ -22,13 +22,14 @@ export default {
 
         $nodes.each( function() {
             const $output = $( this );
+            const output = this;
 
             // nodes are in document order, so we discard any nodes in questions/groups that have a disabled parent
             if ( $output.closest( '.or-branch' ).parent().closest( '.disabled' ).length ) {
                 return;
             }
 
-            const expr = $output.attr( 'data-value' );
+            const expr = output.dataset.value;
             /*
              * Note that in XForms input is the parent of label and in HTML the other way around so an output inside a label
              * should look at the HTML input to determine the context.
@@ -36,31 +37,32 @@ export default {
              * or the parent with a name attribute
              * or the whole document
              */
-            let $context = $output.closest( '.question, .or-group' );
+            let context = output.closest( '.question, .or-group' );
 
-            if ( !$context.is( '.or-group' ) ) {
-                $context = $context.find( '[name]' ).eq( 0 );
+
+            if ( !context.matches( '.or-group' ) ) {
+                context = context.querySelector( '[name]' );
             }
 
-            let context = that.form.input.getName( $context );
+            let contextPath = that.form.input.getName( context );
 
             /* 
              * If the output is part of a group label and that group contains repeats with the same name,
              * but currently has 0 repeats, the context will not be available. See issue 502. 
              * This same logic is applied in branch.js.
              */
-            if ( $context.children( `.or-repeat-info[data-name="${context}"]` ).length && !$context.children( `.or-repeat[name="${context}"]` ).length ) {
-                context = null;
+            if ( $( context ).children( `.or-repeat-info[data-name="${contextPath}"]` ).length && !$( context ).children( `.or-repeat[name="${contextPath}"]` ).length ) {
+                contextPath = null;
             }
 
             const insideRepeat = ( clonedRepeatsPresent && $output.parentsUntil( '.or', '.or-repeat' ).length > 0 );
             const insideRepeatClone = ( insideRepeat && $output.parentsUntil( '.or', '.or-repeat.clone' ).length > 0 );
-            const index = ( insideRepeatClone && context ) ? that.form.input.getIndex( $context ) : 0;
+            const index = ( insideRepeatClone && contextPath ) ? that.form.input.getIndex( context ) : 0;
 
             if ( typeof outputCache[ expr ] !== 'undefined' ) {
                 val = outputCache[ expr ];
             } else {
-                val = that.form.model.evaluate( expr, 'string', context, index, true );
+                val = that.form.model.evaluate( expr, 'string', contextPath, index, true );
                 if ( !insideRepeat ) {
                     outputCache[ expr ] = val;
                 }