+++ src/js/required.js
@@ -20,10 +20,11 @@ export default {
 
         $nodes.each( function() {
             const $input = $( this );
-            const requiredExpr = that.form.input.getRequired( $input );
-            const path = that.form.input.getName( $input );
+            const input = this;
+            const requiredExpr = that.form.input.getRequired( input );
+            const path = that.form.input.getName( input );
             // Minimize index determination because it is expensive.
-            const index = repeatClonesPresent ? that.form.input.getIndex( $input ) : 0;
+            const index = repeatClonesPresent ? that.form.input.getIndex( input ) : 0;
             // The path is stripped of the last nodeName to record the context.
             // This might be dangerous, but until we find a bug, it improves performance a lot in those forms where one group contains
             // many sibling questions that each have the same required expression.