+++ src/js/widget.js
@@ -1,4 +1,3 @@
-import $ from 'jquery';
 import input from './input';
 import event from './event';
 const range = document.createRange();
@@ -131,7 +130,7 @@ class Widget {
      * @memberof Widget
      */
     get originalInputValue() {
-        return input.getVal( $( this.element ) );
+        return input.getVal( this.element );
     }
 
     /**
@@ -141,7 +140,10 @@ class Widget {
      * @memberof Widget
      */
     set originalInputValue( value ) {
-        input.setVal( $( this.element ), value, null );
+        // Avoid unnecessary change events as they could have significant negative consequences!
+        // However, to add a check for this.originalInputValue !== value here would affect performance too much,
+        // so we rely on widget code to only this setter when the value changes.
+        input.setVal( this.element, value, null );
         this.element.dispatchEvent( event.Change() );
     }