+++ src/widget/range/range-widget.js
@@ -34,14 +34,22 @@ class RangeWidget extends Widget {
         }
 
         this.range.addEventListener( 'change', () => {
-            this.current.textContent = this.value;
-            this.originalInputValue = this.value;
-            this._updateMercury( ( this.value - this.props.min ) / ( that.props.max - that.props.min ) );
+            // Avoid unnecessary change events on original input as these can have big negative consequences
+            // https://github.com/OpenClinica/enketo-express-oc/issues/209
+            if ( this.originalInputValue !== this.value ) {
+                this.current.textContent = this.value;
+                this.originalInputValue = this.value;
+                this._updateMercury( ( this.value - this.props.min ) / ( that.props.max - that.props.min ) );
+            }
         } );
 
-        // Do not use change handler for this because this doesn't if the user clicks on the internal DEFAULT
+        // Do not use change handler for this because this doesn't fire if the user clicks on the internal DEFAULT
         // value of the range input.
         this.widget.querySelector( 'input.empty' ).addEventListener( 'click', () => {
+            this.range.classList.remove( 'empty' );
+            this.range.dispatchEvent( events.Change() );
+        } );
+        this.widget.querySelector( 'input.empty' ).addEventListener( 'touchstart', () => {
             this.range.classList.remove( 'empty' );
             this.range.dispatchEvent( events.Change() );
         } );
@@ -113,8 +121,8 @@ class RangeWidget extends Widget {
     _reset() {
         this.value = '';
         this.originalInputValue = '';
-        this.current.textContent = '-';
-        this._updateMercury( -1 );
+        this.current.textContent = '';
+        this._updateMercury( 0 );
     }
 
     disable() {